requirements.txt
fenics_dolfin==2019.1.0
jaxtyping==0.2.28
joblib==1.3.2
matplotlib==3.8.2
numpy==1.24.1
omegaconf==2.3.0
pandas==2.2.1
petsc4py==3.20.2
pytorch_lightning==2.1.3
scipy==1.12.0
seaborn==0.13.2
setuptools==68.2.2
sympy==1.12
sympy==1.11.1
tensordict==0.2.1
torch==2.3.0.dev20231231+cu121
torch_optimizer==0.3.0
torchvision==0.18.0.dev20231231+cu121
tqdm==4.65.0
